# [TR] PIM-5211: display the variant group edit form in less than 2 sec