--- a/src/Pim/Bundle/EnrichBundle/Resources/config/form_subscribers.yml
+++ b/src/Pim/Bundle/EnrichBundle/Resources/config/form_subscribers.yml
@@ -7,6 +7,7 @@ parameters:
     pim_enrich.form.subscriber.set_attribute_group_sort_order.class: Pim\Bundle\EnrichBundle\Form\Subscriber\SetAttributeGroupSortOrderSubscriber
     pim_enrich.form.subscriber.transform_product_template_values.class: Pim\Bundle\EnrichBundle\Form\Subscriber\TransformProductTemplateValuesSubscriber
     pim_enrich.form.subscriber.add_variant_group_template.class: Pim\Bundle\EnrichBundle\Form\Subscriber\AddVariantGroupTemplateSubscriber
+    pim_enrich.form.subscriber.add_variant_group_axes.class: Pim\Bundle\EnrichBundle\Form\Subscriber\AddVariantGroupAxesSubscriber
 
 services:
     # Subscribers
@@ -54,3 +55,8 @@ services:
         class: %pim_enrich.form.subscriber.add_variant_group_template.class%
         arguments:
             - '@pim_user.context.user'
+
+    pim_enrich.form.subscriber.add_variant_group_axes:
+        class: %pim_enrich.form.subscriber.add_variant_group_axes.class%
+        arguments:
+            - %pim_catalog.entity.attribute.class%