Tested on a catalog with 10k attributes. xdebug disabled.
NB: curl calls are slighty faster than real UI calls.

 Mean time of 25 curls call to `/enrich/variant-group/4/edit':
  * before patch: 10.068s (more than 10k requests to DB)
  * after patch: 1.602s (165 requests to DB)

Mean time of 25 curls call to `/enrich/group/5/edit':
  * before patch: 10.983s (more than 10k requests to DB)
  * after patch: 1.237s (160 request to DB)

| Q                 | A
| ----------------- | ---
| Specs             | Y
| Behats            | Y
| Blue CI           | running
| Changelog updated | Y
| Review and 2 GTM  |